python3: fix printing of called box command
To make test box-py/call.test.py compatible with Python 2 and Python 3
it must have the same output when running with both versions of
interpeter. With running on Python 2.7 output of called commands
concluded to round brackets while on Python 3 it is not.

[001] @@ -20,7 +20,7 @@
[001]  - true
[001]  - null
[001]  ...
[001] -call  f1 ()
[001] +('call ', 'f1', ((),))
[001]  - 'testing'
[001]  - 1
[001]  - False

To fix it print should use formatted string in print() call.

Part of #20